requirements.txt
langgraph
langchain-groq
langchain-core
google-search-results
duckduckgo-search
rich
python-dotenv